got PICO working, sending data to the PIC
solution: I checked my wires, then resent the PICO to Serial code, and then it worked. I think I had the wrong version of the UF2 code on the PICO, so make sure to keep the PICO-to-Serial code on the PICO!

+++ hw/hw9/main.c
@@ -50,12 +50,14 @@ int main()
 		{
 			//read encoder value in counts
 			WriteUART2("a");
+
 			while (!get_encoder_flag()) {
 				//delay until encoder/PICO are done sending insructions
 			}
+			
 			set_encoder_flag(0); //prepare for new instructions
 			char m[50];
-			int p = get_encoder_count();
+			float p = get_encoder_count();
 			sprintf(m, "%f \r\n", p);
 			NU32_WriteUART3(m);
 			break;
@@ -65,12 +67,14 @@ int main()
 		{
 			//read encoder value in degrees
 			WriteUART2("a");
+			
 			while (!get_encoder_flag()) {
 				//delay until encoder/PICO are done sending insructions
 			}
+			
 			set_encoder_flag(0); //prepare for new instructions
 			char m[50];
-			int p = get_encoder_count();
+			float p = get_encoder_count();
 			sprintf(m, "%f \r\n", p * 360 /(4*334) );
 			NU32_WriteUART3(m);
 			break;
